// ui/esignature/src/app/softcafe/constants/content-type.enum.ts
export enum ContentType {
  User = "User",
  Role = "Role",
  Branch = "Branch",
  Search = "Search",
  Location = "Location",
  DataConfig = "DataConfig",
  Dashboard = "Dashboard",
  ProcessLog = "ProcessLog",
  Payment = "Payment",
  GenericMap = "GenericMap",
  CustomerMailConfig = "CustomerMailConfig",
  BusinessException = "BusinessException",
  MailTemplete = "MailTemplete",
  FundRequisition = "FundRequisition",
  Customer = "Customer",
  SConfiguration = "SConfiguration",
  CustomerMailConfigDetails = "CustomerMailConfigDetails",
  MtType = "MtType",
  NameVariant = "NameVariant",
  Group = "Group",
  FundSetting = "FundSetting",
  SmsTemplate = "SmsTemplate",
  MailTracker = "MailTracker",
  SmsTracker = "SmsTracker",
  LegalEntity = "LegalEntity",
  RoleGroup = "RoleGroup",
  AppPermission = "AppPermission",
  Signature = 'Signature',
  SignatureInfo = 'SignatureInfo',
  SecurityQuestion = 'SecurityQuestion',
  Signatory = "Signatory",
  SecurityQuestionAnswer = "SecurityQuestionAnswer",
  FreeText = "FreeText",
  Institution = "Institution",
  Request = "Request",
  DocumentFiles = "DocumentFiles",
  ActivityLog = "ActivityLog",
  ReportSearch = "ReportSearch",
}